+++ remora/src/main/java/remora/remora/FrameExtraction/FrameExtractionController.java
@@ -6,7 +6,7 @@ import remora.remora.FrameExtraction.dto.FrameExtractionRequestDto;
 import remora.remora.FrameExtraction.dto.FrameExtractionResponseDto;
 
 @RestController
-public class FrameExtractinController {
+public class FrameExtractionController {
     FrameExtractionService frameExtractionService = new FrameExtractionService();
 
     public FrameExtractionResponseDto frameExtract(FrameExtractionRequestDto request) {
