+++ remora/src/main/java/remora/remora/MainControl/MainController.java
@@ -1,19 +1,30 @@
 package remora.remora.MainControl;
 
 import java.io.File;
+import java.io.IOException;
+import java.util.List;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.*;
 
-import remora.remora.FrameExtraction.FrameExtractinController;
+import org.springframework.web.multipart.MultipartFile;
+import remora.remora.Api.ApiController;
+import remora.remora.Api.ApiDTO.DeleteRequestDto;
+import remora.remora.Api.ApiDTO.SimpleResponseDto;
+import remora.remora.Api.ApiDTO.UploadRequestDto;
+import remora.remora.Api.ApiDTO.UploadResponseDto;
+import remora.remora.FrameExtraction.FrameExtractionController;
 import remora.remora.FrameExtraction.dto.FrameExtractionRequestDto;
 import remora.remora.FrameExtraction.dto.FrameExtractionResponseDto;
 
 @RestController
 public class MainController {
+    FrameExtractionController frameExtractionController = new FrameExtractionController();
+    ApiController apiController = new ApiController();
+
     @GetMapping("/test/extraction")
     public void testExtraction() {
-        FrameExtractinController frameExtractionController = new FrameExtractinController();
+        //FrameExtractinController frameExtractionController = new FrameExtractinController();
         FrameExtractionRequestDto request = new FrameExtractionRequestDto();
         FrameExtractionResponseDto response = null;
 
@@ -22,5 +33,21 @@ public class MainController {
         if(response.success) {
             System.out.println("Success frame extraction");
         }
+    }
+
+    @GetMapping(value = "/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    UploadResponseDto uploadVideo(@RequestParam("originVideo") List<MultipartFile> files,
+                                  @RequestParam("needTranslate") String needTranslate) throws IOException {
+        return apiController.uploadVideo(files, needTranslate);
+    }
+
+    @PutMapping("/upload")
+    SimpleResponseDto changeVideo(@RequestBody UploadRequestDto uploadReqDto) {
+        return apiController.changeVideo(uploadReqDto);
+    }
+
+    @DeleteMapping("/upload")
+    SimpleResponseDto deleteVideo(@RequestBody DeleteRequestDto deleteReqDto){
+        return apiController.deleteVideo(deleteReqDto);
     }
 }
